# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * D&D Adventure Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GameStateStatus =
  (typeof GameStateStatus)[keyof typeof GameStateStatus];

export const GameStateStatus = {
  idle: "idle",
  active: "active",
  paused: "paused",
} as const;

export interface GameState {
  status: GameStateStatus;
  activeCharacterIds: number[];
  /** @nullable */
  storyContext?: string | null;
  /** @nullable */
  dmName?: string | null;
  updatedAt: string;
}

export type NarrativeEntryType =
  (typeof NarrativeEntryType)[keyof typeof NarrativeEntryType];

export const NarrativeEntryType = {
  dm: "dm",
  player: "player",
  system: "system",
} as const;

export interface NarrativeEntry {
  id: number;
  type: NarrativeEntryType;
  /** @nullable */
  username?: string | null;
  content: string;
  createdAt: string;
}

export interface ActionRequest {
  username: string;
  action: string;
}

export type CharacterStatus =
  (typeof CharacterStatus)[keyof typeof CharacterStatus];

export const CharacterStatus = {
  pending: "pending",
  active: "active",
} as const;

export interface Character {
  id: number;
  username: string;
  name: string;
  race: string;
  class: string;
  status: CharacterStatus;
  level: number;
  xp: number;
  hp: number;
  maxHp: number;
  strength: number;
  dexterity: number;
  constitution: number;
  intelligence: number;
  wisdom: number;
  charisma: number;
  /** @nullable */
  subChoice?: string | null;
  inventory: string[];
  skills: string[];
  createdAt: string;
}

export interface CreateCharacterBody {
  username: string;
  name: string;
  race: string;
  class: string;
  subChoice?: string;
}

export interface AwardXpBody {
  amount: number;
}

export interface ErrorResponse {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface DmNameResponse {
  dmName: string;
}

export interface Campaign {
  id: number;
  name: string;
  dmName: string;
  gameStatus: string;
  /** @nullable */
  storyContext?: string | null;
  narrativeJson: string;
  charactersJson: string;
  savedAt: string;
}

export interface SaveCampaignBody {
  name: string;
}