# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: D&D Adventure Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters/create:
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a new pending character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterBody"
      responses:
        "201":
          description: Character created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get character details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Remove a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /characters/{id}/activate:
    post:
      operationId: activateCharacter
      tags: [characters]
      summary: Activate a pending character (add to party)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "400":
          description: Party full or already active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /characters/{id}/deactivate:
    post:
      operationId: deactivateCharacter
      tags: [characters]
      summary: Deactivate an active character (move to pending)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
  /characters/{id}/award-xp:
    post:
      operationId: awardXp
      tags: [characters]
      summary: Award XP to a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwardXpBody"
      responses:
        "200":
          description: XP awarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [game]
      summary: List all saved campaigns
      responses:
        "200":
          description: List of saved campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: saveCampaign
      tags: [game]
      summary: Save the current campaign state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCampaignBody"
      responses:
        "201":
          description: Campaign saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [game]
      summary: Get a specific campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCampaign
      tags: [game]
      summary: Delete a saved campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/load:
    post:
      operationId: loadCampaign
      tags: [game]
      summary: Load a saved campaign (restores full game state)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign loaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameState"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    GameState:
      type: object
      properties:
        status:
          type: string
          enum: [idle, active, paused]
        activeCharacterIds:
          type: array
          items:
            type: integer
        storyContext:
          type: ["string", "null"]
        dmName:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required:
        - status
        - activeCharacterIds
        - updatedAt
...
    Character:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        race:
          type: string
        class:
          type: string
        status:
          type: string
          enum: [pending, active]
        level:
          type: integer
        xp:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        strength:
          type: integer
        dexterity:
          type: integer
        constitution:
          type: integer
        intelligence:
          type: integer
        wisdom:
          type: integer
        charisma:
          type: integer
        subChoice:
          type: ["string", "null"]
        inventory:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - name
        - race
        - class
        - status
        - level
        - xp
        - hp
        - maxHp
        - strength
        - dexterity
        - constitution
        - intelligence
        - wisdom
        - charisma
        - inventory
        - skills
        - createdAt
    CreateCharacterBody:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        race:
          type: string
        class:
          type: string
        subChoice:
          type: string
      required:
        - username
        - name
        - race
        - class
    AwardXpBody:
      type: object
      properties:
        amount:
          type: integer
      required:
        - amount
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dmName:
          type: string
        gameStatus:
          type: string
        storyContext:
          type: ["string", "null"]
        narrativeJson:
          type: string
        charactersJson:
          type: string
        savedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - dmName
        - gameStatus
        - narrativeJson
        - charactersJson
        - savedAt
    SaveCampaignBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name