# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * D&D Adventure Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { CharacterStatus } from "./characterStatus";

export interface Character {
  id: number;
  username: string;
  name: string;
  race: string;
  class: string;
  status: CharacterStatus;
  level: number;
  xp: number;
  hp: number;
  maxHp: number;
  strength: number;
  dexterity: number;
  constitution: number;
  intelligence: number;
  wisdom: number;
  charisma: number;
  /** @nullable */
  subChoice?: string | null;
  inventory: string[];
  skills: string[];
  createdAt: Date;
}