# lib/api-zod/src/generated/types/createCharacterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * D&D Adventure Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCharacterBody {
  username: string;
  name: string;
  race: string;
  class: string;
  subChoice?: string;
}