# lib/api-zod/src/generated/types/gameState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * D&D Adventure Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { GameStateStatus } from "./gameStateStatus";

export interface GameState {
  status: GameStateStatus;
  activeCharacterIds: number[];
  /** @nullable */
  storyContext?: string | null;
  /** @nullable */
  dmName?: string | null;
  updatedAt: Date;
}